GUAC-1344: Add TLS-PSK support (RFC 4279)
This allows for mutual authentication with pre-shared keys instead
of certificates (PSK and DHE-PSK) or in addition to regular TLS with
certificates (RSA-PSK).

Adds configuration entry 'psk_peer' (ssl) and corresponding '-P' command
line argument which take a PSK peer in the form of [identity:key].

Multiple occurrences of PSK peers are allowed (the key is searched
according to the identity provided by the client).

Signed-off-by: Isaac Boukris <iboukris@gmail.com>